Move test helper files to `lib/helpers`
We want to have the test helpers we've defined in `tapioca` be available
for extensions/plugins for `tapioca`. This means having our built-in
test helpers available outside of the test directory.

See also: #94
